Updated contactusRouter with administrative privileges on GET and DELETE HTTP methods

+++ routes/contactusRouter.js
@@ -29,14 +29,19 @@ contactusRouter.route('/')
 
 contactusRouter.route('/visitors')
 .get(authenticate.verifyUser, (req, res, next) => {
-    Visitor.find()
-    .then(visitors =>{
-      res.statusCode = 200;
-      res.setHeader('Content-Type', 'application/json');
-      res.json(visitors);
-    })
-    .catch(err => next(err));
-      
+  if(!req.user.admin){
+    res.statusCode = 401;
+    res.setHeader('Content-Type', 'application/json');
+    res.end('You are not authorized to access records.');
+  } else if(req.user.admin){
+      Visitor.find()
+      .then(visitors =>{
+        res.statusCode = 200;
+        res.setHeader('Content-Type', 'application/json');
+        res.json(visitors);
+      })
+      .catch(err => next(err));
+  }
 })
 .post((req, res) =>{
   res.statusCode = 403;
@@ -46,25 +51,37 @@ contactusRouter.route('/visitors')
   res.statusCode = 403;
   res.end('PUT operation is not supported on /contactus/visitors');
 })
-.delete((req, res, next) =>{
-    Visitor.deleteMany()
-    .then(response => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(response);
-    })
-    .catch(err =>next(err))
+.delete(authenticate.verifyUser, (req, res, next) =>{
+  if(!req.user.admin){
+    res.statusCode = 401;
+    res.setHeader('Content-Type', 'application/json');
+    res.end('You are not authorized to delete records.');
+  } else if(req.user.admin){
+      Visitor.deleteMany()
+      .then(response => {
+          res.statusCode = 200;
+          res.setHeader('Content-Type', 'application/json');
+          res.json(response);
+      })
+      .catch(err =>next(err));
+    }
   });
 
   contactusRouter.route('/visitors/:visitorId')
   .get(authenticate.verifyUser, (req, res, next) => {
-      Visitor.findById(req.params.visitorId)
-      .then(visitor =>{
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(visitor);
-      })
-      .catch(err => next(err));
+    if(!req.user.admin){
+      res.statusCode = 401;
+      res.setHeader('Content-Type', 'application/json');
+      res.end('You are not authorized to access this record.');
+    } else if(req.user.admin){
+        Visitor.findById(req.params.visitorId)
+        .then(visitor =>{
+          res.statusCode = 200;
+          res.setHeader('Content-Type', 'application/json');
+          res.json(visitor);
+        })
+        .catch(err => next(err));
+      }
         
   })
   .post((req, res) =>{
@@ -75,14 +92,20 @@ contactusRouter.route('/visitors')
     res.statusCode = 403;
     res.end(`PUT operation is not supported on /contactus/visitors/${req.params.visitorId}/`);
   })
-  .delete((req, res, next) =>{
-      Visitor.findByIdAndDelete(req.params.visitorId)
-      .then(response => {
-          res.statusCode = 200;
-          res.setHeader('Content-Type', 'application/json');
-          res.json(response);
-      })
-      .catch(err =>next(err))
+  .delete(authenticate.verifyUser, (req, res, next) =>{
+    if(!req.user.admin){
+      res.statusCode = 401;
+      res.setHeader('Content-Type', 'application/json');
+      res.end('You are not authorized to delete record.');
+    } else if(req.user.admin){
+        Visitor.findByIdAndDelete(req.params.visitorId)
+        .then(response => {
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(response);
+        })
+        .catch(err =>next(err));
+      }
     });
 
 module.exports = contactusRouter;
